refactor: split create/edit opportunity modal into steps with RHF+zod (Pitch 2)
Per #676's approved "Pitch 2", restructures CreateVolunteerOpportunityModal
from a single ~1400-line file into per-step components sharing a
react-hook-form + zod form (the first form library used in the codebase),
and closes several UX/a11y gaps:

- Plain white header (drop the one-off gradient accent bar) matching the
  rest of the app's modals; the stepper moves into its own row in the body.
- Per-step validation on "Next"/stepper-forward-jump instead of only at
  Publish time - a required field left empty is caught before the user
  reaches the end of the wizard.
- aria-required/aria-describedby wiring so required fields and their error
  text reach assistive tech, a focus trap on open, and an aria-live region
  plus step-heading focus announcing step changes.
- Responsive grids (street/number, zip/city, radio-card groups) that no
  longer force two columns on narrow viewports.
- Tags are now a chip/pill input (Enter or comma to add, backspace to
  remove the last, case-insensitive dedupe) instead of raw comma-separated
  text.
- An unsaved-changes guard: closing a dirty form (backdrop/Escape/close
  button) asks for confirmation instead of silently discarding input.

Addresses #676
