+++ frontend/package.json
@@ -17,6 +17,7 @@
     "i18n:check": "node scripts/check-i18n-keys.js"
   },
   "dependencies": {
+    "@hookform/resolvers": "^5.4.0",
     "date-fns": "4.4.0",
     "i18next": "26.3.6",
     "i18next-browser-languagedetector": "8.2.1",
@@ -26,10 +27,12 @@
     "react": "19.2.7",
     "react-big-calendar": "1.20.0",
     "react-dom": "19.2.7",
+    "react-hook-form": "^7.81.0",
     "react-i18next": "17.0.9",
     "react-leaflet": "5.0.0",
     "react-oidc-context": "3.3.1",
-    "react-router": "8.2.0"
+    "react-router": "8.2.0",
+    "zod": "^4.4.3"
   },
   "devDependencies": {
     "@eslint/js": "10.0.1",
